+++ config/urls.py
@@ -13,4 +13,5 @@ urlpatterns = [
     path('api/review/', include('apps.api.urls.review_urls', namespace='review_endpoints')),
     path('api/review_settings/', include('apps.api.urls.review_settings_urls', namespace='review_settings_endpoints')),
     path('api/answer/', include('apps.api.urls.answer_urls', namespace='answer_endpoints')),
+    path('api/qrcode/', include('apps.api.urls.qrcode_urls', namespace='qrcode_endpoints')),
 ]